requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
pymongo==4.4.1
python-dotenv==1.0.0